PostgreSQL: Group summaries: Fix type resolution for UNION
Type resolution for UNION (CASE and related constructs) in PostgreSQL
works by taking the UNIONed queries 2 at a time, i.e. the first TWO
queries are resolved, then it takes the result and resolves it with the
next query and so on.

Previously, when using Hostgroup as the first query and Host (or
Service) as the next query, PostgreSQL encounters two NULL values for
the host_handled (or service_handled) column, resulting in the type
text. This leads to the following error:

UNION types text and boolenum cannot be matched.

Now Hostgroup is the last query that ensures there are not two
consecutive NULL values for host_handled (or service_handled).

NOTE: BREAKING: This change breaks other code that extracts the "base
query" which depends on Hostgroup or Servicegroup being the first query
in the getUnions() array, which has yet to be fixed.  In preparation for
a possible fix that could simply use the last query instead of the
first, the Servicegroup query has also been moved to the end.