Fix empty host/service group details when restrictions are active
With recent changes in IPL - ORM [1], utilized relations no longer
perform subquery filters, which is fine since the only designed
difference is that columns from utilized relations are not appended to
the select list. However, this change produces empty results here
because the filter should actually be run in a subquery. We called
utilize() here because we modified the base query to apply the group ID
filters. Now that we're using filters that automatically handle
relations, it's safe to remove the utilize() calls.

[1] https://github.com/Icinga/ipl-orm/pull/44