Avoid duplicate column suggestions in search bar
`ObjectSuggestions::collectFilterColumns()` returns duplicate column names due to the sub-relations of the joined models.
Hence, `ObjectSuggestions::fetchColumnSuggestions()` returns duplicate suggestions for column names for the search bar.
This fix solves this issue by resolving the redefinition of the method parameter `$model` in `ObjectSuggestions::collectFilterColumns()`
and not registering the duplicate models in `ObjectSuggestions::collectRelations()`.