Introduce new restriction `icingadb/scope/commands`
I thought this is worthwhile to have. It allows to limit users
with a specific set of granted command permissions to certain
hosts and services. The difference to the other restrictions
is though, that it doesn't affect the visiblity of objects but
rather the validity of the granted command permissions only.

This is the same effect that the current restrictions already
have. But this allows to set up a role that doesn't limit a user
in what's shown and what's not, but where which commands are
allowed.