Don't strictly check whether `hGet` returns false
Nowadays hGet returns null instead of false if there is no data available yet.
But since we are performing strict comparison on false, it won`t enter the if branch anymore.
There is another Predis client method we use which is `hmGet`, but as usual it still returns only an array.