UnreachableParent: Match unreachable services at the root
A service that has no parent but is unreachable, has a host
which marks it as such. In other words, an implicit dependency.

As long as we don't fully support implicit dependencies,
such a service must be shown as root problem. Otherwise,
the message about insufficient access rights is shown.