ObjectSuggestion: Dynamically adjust columns and custom var suggestion limits (#1173)
Currently, the limit is globally set to 50, hence the short custom
variable names which overlaps with most of the column names will not be
suggested and instead only column names will be suggested. To prevent
this the default limit is split in half and adjusted dynamically based
on available suggestions.

resolves #1094